+++ Packages/Tracking/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Disabled incorrect version check for multiple device mode
 - Fixed instance where setting CapsuleHand colour through code would report a nullref
 - Fixed references to LeapProvider and Camera in PointerElement to prevent nullrefs when changing either in the main UIInputModule
+- Fixed GrabBall for new physical setup
 
 
 ## [7.0.0] - 22/07/2024